Remove Duplicate EVNTs
The duplicate EVNTs created by ANCS that are based on another items (or that are used by more than one model/character set) were causing dependency problems with the patcher because it creates a new id for the same EVNT, and lists this new ID as a dependency.  ANIM dependency is the same ANIM file, but the ANIM can only have a single EVNT id reference.  The ANCS dependency list was including an EVNT that was not present in the ANIM it depended on and thus causing it to crash because it was not building it with the correct EVNT.  This replaces the EVNT dependencies with the EVNTs from the ANIM file.

My code is garbage and likely not very efficient, feel free to correct it.