Merge in code from rehashedsalt/randovania-appimage
This does a lot more than that, but it's the jyst of it, more or less.
The repository in question does not share history with Randovania, so
this is a historyless merge.

It's either this or give Randovania multiple origin commits, and I'd
rather not have to put anybody through that.