Dread: Add minor option in Ghavoran Total Recharge Station North
Allow traveral between the upper and lower parts of this room without
Morph Ball or sliding, by destroying the left Enky, which opens the path
until the Grapple Box gets pulled.

Note: This is normally not going to come into play, since you normally
need to have Morph Ball to arrive here regardless of the path. But it could
be relevant with a starting position inside Ghavoran.