Introduce template for normal bombs
The Lay Bomb template can use either the normal Bomb or the Cross Bomb. In
most cases Cross Bombs can do all the same things that normal Bombs can do,
however, for diagonal bomb jumps can't really be done with Cross Bombs, so
in order to use a nice template to get shorter syntax in the logic
expressions, this template is introduced.