Dread: Better handling of Chozo X Trigger in Ghavoran Spin Boost Room
Reorganize the connections in Spin Boost Room so that the event is required
on the way out rather than on the way in. Add requirement to either use
Highly Dangerous logic or already having the means to fight the Chozo X
when leaving the room.

This change is meant to help the solver get through more seeds.