Resolver: Only delay recursing into dangerous events
Dread has some items as dangerous resources, but it's suboptimal to avoid
them because the situations in which they are dangerous are higly specific.

It's considered more likely that these items are the required path forward
than the path where we require not having them to be the required path
forward.