Dread late game fixes (#4223)
* Dread: Move requirement to release the X to fight Experiment Z-57

The resolver appears to not understand that it needs to take this
connection after releasing the X, sometimes. If it just doesn't understand
that it needs to use the bomb launcher after having released the X when it
has used it previously, then this is a simple fix.

Technically speaking, this bomb launcher doesn't need either of these two
events, but if the thermal device hasn't been used, then taking the
launcher leads to a softlock anyway. And after using the thermal device, it
is trival to move into the Z-57 room anyway.

* Dread: Add missing requirements when entering Elun Chozo X from the left

* Dread: Add missing requirement to release the X before leaving Elun

---------

Co-authored-by: dyceron <38679103+dyceron@users.noreply.github.com>