Rename "lounge" to "Lounge" in Cave Story database
Fixes an issue where "Lounge.tsc" cannot be found in Linux builds of
Randovania. This is because the filename is written in capital case
whereas it is referenced in this database as lowercase, causing patching
to fail. It also fixes an adjacent issue where editing this area in the
database GUI would crash the application (I assume because it could not
find the image corresponding to that room).

The test cases have been updated as there were two room entries: one for
"Lounge" and one for "lounge". These have been merged into the one entry
with the combined pickups and music defined.