Dread: Switch which shinespark in Gravity Suit Tower requires Power Bombs
The shinespark in the lower part doesn't need powerbombs, because it is
implicilty required to reach the event in Gravity Suit Room. Meanwhile, the
one to reach the top of the room requires power bombs when door lock rando
is disabled and you have no missiles, which means you can't get to the
event in the upper part of Gravity Suit Room.