Dread/burenia shinesparks (#4270)
* Dread: Add trickless shinespark in Main Hub Tower Bottom with Gravity Suit

* Dread: Add shinespark without Gravity Suit in Main Hub Tower Bottom

* Dread: Add shinespark with gravity at the bottom of Garvity Suit Tower

* Dread: Cross gap in Burenia Main Hub Tower Middle with Spin Boost

* Dread: Reach Magnet Wall in Burenia Main Hub Tower Middle with Shinespark

* Dread: Climb up to the charge beam door in Main Hub Tower Middle

* Dread: Climb from Charge Beam Door in Main Hub Tower Middle

* Dread: Move climb to charge beam door to start from the tunnel

* Dread: Allow shinespark from upper Gravity Suit Room outside lock rando

* Dread: Adjust trick level for crossing gap in Main Hub Tower Middle

* Dread: Fix going through upper tunnel to Green TP before breaking blob

* Dread: Switch which shinespark in Gravity Suit Tower requires Power Bombs

The shinespark in the lower part doesn't need powerbombs, because it is
implicilty required to reach the event in Gravity Suit Room. Meanwhile, the
one to reach the top of the room requires power bombs when door lock rando
is disabled and you have no missiles, which means you can't get to the
event in the upper part of Gravity Suit Room.

* Dread: Convert underwater tunnel to Slide Tunnel

* Fix changelog order

Apparently a merge conflict was resolved in a way that put these in the
wrong order.