resolved: Allow negated requirements
To properly avoid repeating an area with a known softlock, we need all
non-satisdfied actions to be identified. If we don't, then the entire
requirement set must be simplied to trivial.

While it's impossible to "collect" a negated resource, having it as an
additional requirment is valuable, because we can avoid it until we drop
the negated resource by rolling back.