Dread: Make Drogyga safe
Jumping up from the pickup doesn't need anything else than Morph Ball
because there is no water in that part of the tunnel.

Adding the Highly Dangerous Logic condition on the last usage of "Before
Drogyga" makes it so Drogyga is only a dangerous resource when playing with
Highly Dangerous Logic enabled.