Dread: Add Highly Dangerous Logic to Grapple Movement in Burenia Main Hub
This connection makes lowering the Spider Magnet Wall a dangerous action.
However, this exact connection is currently never relevant, because you
can always reach the platform from above, using the same tools. If you
could start in this room or in the lower part of the room to the right,
then this could be relevant.

It is probably more valuable to make this action not dangerous. By adding
the Highly Dangerous Logic constraint, the action is only considered
dangerous when this mode is enabled.