Resolver: Avoid using damage requirements as additional in problematic case
When a damage requirement is not satisfied, but it is satisfied with the
full health capacity, the requirement can't be used as an additional. The
problem is that the connection into this action can have a damage
requirement, and we don't have a good way to add that damage to the node
that is getting additional requirement set on it.