Combine damage reductions with min function instead of multiplication
The damage reductions from suits either completely mitigate damage, or are
encoded as the absolute damage multiplier. In cases where multiple suits
provide a non-zero damage multiplier, multiplying together implies that
they are programmed as having a multiplicative effect that stacks.
However, they are implemented as the absolute multiplier.

The only game with a different behaviour is Prime, which handles this
separately in the bootstrap.