Fix missed detection of energy as an intersting resource
With the assumption that the requirement lists that get passed to this
function are unsatisfied, it seems like an anomaly to check whether each
list is satisfied or not. If a list is considered satisfied, but contains a
damage resource, by assuming that the requirement list was unsatisfied due
to energy being less than max capacity due to perious requirements, we can
yield resources for energy as interesting resources.