Blank Dev Game: Add Heated Room and Health Pickup
Add A heated room, with a pickup node. Connections between the door and the
pickup both require damage. The values are chosen such that the starting
health is enough to reach the pickup, but not enough to get back to the
door, on all damage strictness levels. The Health Pickup is added to the
pool, and the health capacity gained from it is upped to 100, to work with
the intended effect on the heated room.