Resolver: Fix events and node resources as additional resources
Fix the catch 22 where these kinds of resources can't be additionals
because they are required to be satisfiable resources for the purpose of
finding the interesting resources.

The solution is to double store two sets of satisfiable resources in the
reach, one that includes the requirement to leave the current node and one
without.