Resolver: Ignore satisfied requirements in additionals
This function is only used to set additional requirents. In this context
trivial requirements should be ignored, because the only reasons left why
there would be trivial requirements here are when only only accessible
resources are skipped because they are either not interesting or have
unsatisfied additional requirements. In both of these cases it's safe to
ignore the satisfied requirements.