Dread: Add conditions to get away from the door into Dairon Freezer
Having these conditions on leaving the door coming into Dairon Freezer from
the left ensures that the logic machinery sees the requirements to navigate
the cold room before it sees the branch that enables navigating through
without having turned on the generator. This is imporant for situations
where the resolver has to roll back and set approproate requirements on
exploring the same paths again. Otherwise it would risk exploring this path
over and over again with the only difference being whether the generator is
enabled or not.