Dread: Remove dangerous requirement on entering Artaria Central Unit
It's possible to enter the Central Unit room from above also after having
beaten White EMMI. Having this connection is relevant to let the logic
machinery access the dock nodes. When the resolver has to do rollbacks,
seeing both conditions that require the central unit event to be collected
and not has a risk of causing repeated unfruitful attempts at collecting
the same resource.