Blank Dev Game: Add room with contraption for progressive items
Add Ledge Room, a room where you have access to one pickup immediately, but
to get back out you need the double jump. The bottom of the room also gives
access to another pickup when you have the jump.

This setup is designed to reproduce a situation where you get access to
the first item in a progressive chain inside of a point of no return, where
you also need the second item in the progressive chain to escape.