Resolver: Fix using the wrong state when finding the resources gained
These resources are meant to be be the resources gained by collecting the
node that's being rolled back. Given that the resource can depends on the
current inventory, it must use the state from before collecting the node.

Particularly in the case of progressive items, it matters that we pass the
context from before collecting the node into this function.

These resources are used to detect the point at which we roll back
collecting a dangerous resource that where we see the requirement that
makes it dangerous. If any games would have a chain of progressive items
where any of the items in the chain were dangerous, then this fix would
likely be important.