Dread: Make event nodes for Artaria Single Use Slide accessible later
Having the event node not be accessible after doing the dangerous event
leads to the resolver seeing this resource as something that unlocks more
of the world when it gets rolled back. Any time the resolver is stuck, it
will add this requirement on all the nodes it rolls back, and it will
reexplore them once the dangerous event has been rolled back.