Dread: Ignore cold requirements for right side of Dairon Freezer
Make Dairon Power Switch 2 not dangerous in presets that don't use Highly
Dangerous Logic. Hand wave away the dangerous nature of navigating through
the right side of the room. Aside from the obvious benefit of making the
generator more likely to activate the generator, this will also make the
resolver both activate it quicker and will also avoid problems where
rolling it back causes unnecessary actions to be retried. In other words,
seeds requiring deep rollbacks benefit from this family of changes.

This is a hand wave, but the idea is that there's nothing important that
appears to be logical in this section of the Freezer after having turned
on the power.