Generator: Move check if beatable check to after assigning pool results
Games with pre placed key items such as Dread need these items to be
accessible in order to beat the game. Considering that these are placed
without checking if the locations are accessible, it stands to reason that
the check_if_beatable logic should run after these are placed, to reject
base patches that make these key items inaccessible even when having
everything else in the item pool.