Videos added, SEC2 trick reversed
Videos added for the new logic. Hypermode wall jump has been removed from Security Room. Upon research, with single wall jumps there is a 25 frame window. With the physics as it currently is for wall jumping, it is just a 5 frame window and might be considerably too tight to be considered humanly possible.