Log bug fixes for Sector 4
Corrected powamp shaft node at the top of the room to require standing on frozen enemies or space jump to get up to the door or sector connector. Added damage run logic to the drain pipe from the item for consistency. Added shinespark trick to get through Powamp in aquarium hub.