Generator: Try harder to check if unsafe resource is safe
Try to collect all safe resources after an unsafe action before checking if
the initial node is reachable. This enables cases like: door lock is past
a point of no return, but behind that door is other doors and an event, and
collecting all of them opens up the path back.