Generator: Fix not filtering paths where the initial node is unreachable
When inside a path of no return, there might be a lot of requirements from
the rest of the world that have been discovered before. Some of those
requirements might be subsets of the requirements that must be satisfied
to do anything. When creating a requirement set for the unsatisfied
requirements, some important requirements might be filtered out. However,
when evaluating actions, it's important that the requirements are full
enough to actually get anywhere.