Generator: Avoid redundant sets of pickups
Remove tuples of pickup entries from the result if there is a different
tuple that is a proper subset.

For most games this is already handled at the item requirement stage, but
when reasoning about the combination of Energy Parts and Energy Tanks and
how they interact with the resource collection in the middle of generation,
it becomes relevant to prune redundancies again here.