keybearer hints choose a random pickup feature
quick and dirty. if this feels better to people we can begin the tuning phase. gaussian params can change as always, and I would like to explore bringing back "missile-related" and the like for these if there's a way we can keep it out of the pool for translator hints